requirements.txt
alembic==1.0.0
astroid==2.0.1
autopep8==1.3.5
Babel==2.6.0
beautifulsoup4==4.6.0
blinker==1.4
click==6.7
Django==2.0.7
django-appconf==1.0.2
django-super-inlines==0.1.4
django-user-accounts==2.0.3
et-xmlfile==1.0.1
extras==1.0.0
fixtures==3.0.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-BabelEx==0.9.3
Flask-Gravatar==0.5.0
Flask-HTMLmin==1.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.11.3
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
linecache2==1.0.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
num2words==0.5.7
olefile==0.45.1
openpyxl==2.5.4
passlib==1.7.1
pbr==4.1.1
pdfkit==0.6.1
pep8==1.7.1
Pillow==5.2.0
pisa==3.0.33
pkg-resources==0.0.0
pycodestyle==2.4.0
pycrypto==2.6.1
pylint==2.0.0
pylint-django==0.11
pylint-plugin-utils==0.2.6
PyPDF2==1.26.0
pyrsistent==0.14.4
python-dateutil==2.7.3
python-editor==1.0.3
python-mimeparse==1.6.0
pytz==2018.5
reportlab==3.5.1
simplejson==3.16.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.10
sqlparse==0.2.4
testtools==2.3.0
traceback2==1.4.0
typed-ast==1.1.0
unittest2==1.1.0
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xhtml2pdf==0.2.2
xlrd==1.1.0
